LayoutTests/fast/canvas/webgl/webgl2-large-getbuffersubdata-expected.txt

Make sure that reading large buffers with getBufferData works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS successfullyParsed is true

TEST COMPLETE
